chore: change aas data plane groupId
