--- data-plane-aas/src/test/java/de/fraunhofer/iosb/DataPlaneAasExtensionTest.java
@@ -1,48 +1,0 @@
-/*
- * Copyright (c) 2021 Fraunhofer IOSB, eine rechtlich nicht selbstaendige
- * Einrichtung der Fraunhofer-Gesellschaft zur Foerderung der angewandten
- * Forschung e.V.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *     http://www.apache.org/licenses/LICENSE-2.0
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package de.fraunhofer.iosb;
-
-import dev.failsafe.RetryPolicy;
-import okhttp3.OkHttpClient;
-import org.eclipse.edc.boot.system.injection.ObjectFactory;
-import org.eclipse.edc.connector.dataplane.spi.pipeline.PipelineService;
-import org.eclipse.edc.junit.extensions.DependencyInjectionExtension;
-import org.eclipse.edc.spi.system.ServiceExtensionContext;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-
-import static org.mockito.Mockito.mock;
-
-@ExtendWith(DependencyInjectionExtension.class)
-class DataPlaneAasExtensionTest {
-
-    private DataPlaneAasExtension extension;
-
-    @BeforeEach
-    void setUp(ServiceExtensionContext context, ObjectFactory factory) {
-        context.registerService(PipelineService.class, mock(PipelineService.class));
-        context.registerService(OkHttpClient.class, mock(OkHttpClient.class));
-        context.registerService(RetryPolicy.class, RetryPolicy.ofDefaults());
-
-        extension = factory.constructInstance(DataPlaneAasExtension.class);
-    }
-
-    @Test
-    void testInitialize(ServiceExtensionContext context) {
-        extension.initialize(context);
-    }
-}
