+++ data-plane-aas/src/test/java/de/fraunhofer/iosb/ilt/aas/lib/http/HttpClientProviderTest.java
@@ -1,0 +1,30 @@
+/*
+ * Copyright (c) 2021 Fraunhofer IOSB, eine rechtlich nicht selbstaendige
+ * Einrichtung der Fraunhofer-Gesellschaft zur Foerderung der angewandten
+ * Forschung e.V.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package de.fraunhofer.iosb.ilt.aas.lib.http;
+
+import de.fraunhofer.iosb.ilt.testutils.TestUtils;
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+
+class HttpClientProviderTest {
+
+    @Test
+    void testClientFor() {
+        assertNotNull(HttpClientProvider.clientFor(new TestUtils().getSelfSignedCertificate()));
+    }
+
+}
