+++ data-plane-aas/src/test/java/de/fraunhofer/iosb/ilt/ssl/impl/NoOpSelfSignedCertificateRetrieverTest.java
@@ -1,0 +1,36 @@
+/*
+ * Copyright (c) 2021 Fraunhofer IOSB, eine rechtlich nicht selbstaendige
+ * Einrichtung der Fraunhofer-Gesellschaft zur Foerderung der angewandten
+ * Forschung e.V.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package de.fraunhofer.iosb.ilt.ssl.impl;
+
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.Test;
+
+import java.net.MalformedURLException;
+import java.net.URL;
+
+import static org.junit.jupiter.api.Assertions.assertTrue;
+
+class NoOpSelfSignedCertificateRetrieverTest {
+
+    @Test
+    void testGetSelfSignedCertificate() throws MalformedURLException {
+        var testSubject = new NoOpSelfSignedCertificateRetriever();
+        var result = testSubject.getSelfSignedCertificate(new URL("https://example.com"))
+                .onSuccess(Assertions::assertNull);
+        // .onFailure(fail()) did not work
+        assertTrue(result.failed());
+    }
+}
